refactor: split ToRelativePath into platform-specific files with build tags
NormalizeMSYSPath had no runtime.GOOS guard, corrupting valid Unix
paths on macOS/Linux (e.g., /tmp/repo/file.txt → /var/folders/...).

Move ToRelativePath and NormalizeMSYSPath into relative_windows.go
(with MSYS normalization + Unix path filter) and relative_unix.go
(plain filepath.Rel, no normalization). Tests split accordingly.

Entire-Checkpoint: cbe58c84ed5b
