+++ cmd/entire/cli/paths/paths.go
@@ -152,53 +152,6 @@ func IsSubpath(parent, child string) bool {
 	return rel != ".." && !strings.HasPrefix(rel, ".."+string(filepath.Separator))
 }
 
-// ToRelativePath converts an absolute path to relative.
-// Returns empty string if the path is outside the working directory.
-func ToRelativePath(absPath, cwd string) string {
-	absPath = NormalizeMSYSPath(absPath)
-
-	// After MSYS normalization, a path starting with "/" that the OS still
-	// doesn't recognize as absolute is an unconvertible Unix path (e.g.,
-	// /home/user/... from a container/sandbox on Windows). Filter it out.
-	if strings.HasPrefix(absPath, "/") && !filepath.IsAbs(absPath) {
-		return ""
-	}
-
-	if !filepath.IsAbs(absPath) {
-		return absPath
-	}
-	relPath, err := filepath.Rel(cwd, absPath)
-	if err != nil || strings.HasPrefix(relPath, "..") {
-		return ""
-	}
-
-	return relPath
-}
-
-// msysDrivePrefix matches MSYS/Git-Bash-style absolute paths like /c/ or /D/.
-// Git for Windows executes hooks through MSYS2 bash, which converts Windows paths
-// (C:\Users\...) to Unix-style (/c/Users/...) in tool output and transcripts.
-var msysDrivePrefix = regexp.MustCompile(`^/([a-zA-Z])/`)
-
-// NormalizeMSYSPath converts MSYS/Git-Bash paths to Windows paths.
-// Handles two MSYS conventions:
-//   - Drive paths: /c/Users/... → C:/Users/...
-//   - Virtual dirs: /tmp/... → <TEMP>/... (MSYS2 maps /tmp to the Windows temp dir)
-//
-// Returns the input unchanged on non-Windows or if the path doesn't match.
-func NormalizeMSYSPath(p string) string {
-	if m := msysDrivePrefix.FindStringSubmatch(p); m != nil {
-		return strings.ToUpper(m[1]) + ":/" + p[3:]
-	}
-	// MSYS2 maps /tmp to the Windows temp directory.
-	if strings.HasPrefix(p, "/tmp/") {
-		if tmp := os.TempDir(); tmp != "" {
-			return filepath.Join(tmp, p[5:])
-		}
-	}
-	return p
-}
-
 // nonAlphanumericRegex matches any non-alphanumeric character
 var nonAlphanumericRegex = regexp.MustCompile(`[^a-zA-Z0-9]`)
