+++ cmd/entire/cli/paths/relative_unix.go
@@ -1,0 +1,22 @@
+//go:build unix
+
+package paths
+
+import (
+	"path/filepath"
+	"strings"
+)
+
+// ToRelativePath converts an absolute path to relative.
+// Returns empty string if the path is outside the working directory.
+func ToRelativePath(absPath, cwd string) string {
+	if !filepath.IsAbs(absPath) {
+		return absPath
+	}
+	relPath, err := filepath.Rel(cwd, absPath)
+	if err != nil || strings.HasPrefix(relPath, "..") {
+		return ""
+	}
+
+	return relPath
+}
